requirements-test.txt
-r requirements.txt

pytest-spark==0.6.0
pytest-mock==3.5.1
pytest-helpers-namespace==2019.1.8
pytest-env==0.6.2
awsglue3==0.0.1
pytest==6.2.2
black~=20.8b1